backend/recipes_kitchen/requirements.txt
asgiref==3.11.0
attrs==25.4.0
black==26.1.0
click==8.3.1
Django==6.0.1
django-crispy-forms==2.5
django-environ==0.12.1
django-filter==25.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==25.1.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.0
platformdirs==4.5.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
pytokens==0.4.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
uritemplate==4.2.0
